requirements.txt
annotated-types==0.7.0
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
dj-database-url==3.0.1
django==4.2.27
django-ninja==1.5.1
idna==3.11
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-core==2.41.5
python-slugify==8.0.4
requests==2.32.5
sqlparse==0.5.5
text-unidecode==1.3
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.2
